backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
aiomqtt==2.1.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
asyncpg==0.29.0
alembic==1.13.2
pyjwt==2.9.0
argon2-cffi==23.1.0
python-multipart==0.0.9
